# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "QR Code Link Generator API",
        "description": "API documentation for the QR Code Link Generator application.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/analytics/overview": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Get an overview of all link analytics for the user",
                "operationId": "256f404efb9fb5047e1ff37c5b8ffc25",
                "responses": {
                    "200": {
                        "description": "Analytics overview",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_clicks": {
                                            "type": "integer",
                                            "example": 1540
                                        },
                                        "total_links": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "top_links": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/links/{id}/analytics": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Get detailed analytics for a specific link",
                "operationId": "9fca3604901e25849c56ff2a85bd4e89",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Link analytics data"
                    },
                    "404": {
                        "description": "Link not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/links": {
            "get": {
                "tags": [
                    "Links"
                ],
                "summary": "Get all links for the authenticated user",
                "operationId": "a4cdfccec2072bff462ecc4a12651ca1",
                "responses": {
                    "200": {
                        "description": "List of links",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Links"
                ],
                "summary": "Create a new short link",
                "operationId": "473714d8872dd56ff0565acb7d029ec6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "original_url"
                                ],
                                "properties": {
                                    "original_url": {
                                        "type": "string",
                                        "format": "uri",
                                        "example": "https://google.com"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Google Search"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Link created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/links/{id}": {
            "get": {
                "tags": [
                    "Links"
                ],
                "summary": "Get a specific link",
                "operationId": "1140569bb06802b6bda1e8aa3f5f8c89",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Link details"
                    },
                    "404": {
                        "description": "Link not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Links"
                ],
                "summary": "Update an existing link",
                "operationId": "e26b4a1a2fb44a2e0e00ec2aaa28c956",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "original_url": {
                                        "type": "string",
                                        "format": "uri"
                                    },
                                    "title": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Link updated"
                    },
                    "404": {
                        "description": "Link not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Links"
                ],
                "summary": "Delete a link",
                "operationId": "8d796148b9d641c83be0a2dce357936d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Link deleted"
                    },
                    "404": {
                        "description": "Link not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/links/{id}/qr-branding": {
            "post": {
                "tags": [
                    "Links"
                ],
                "summary": "Update QR code branding",
                "operationId": "0fc8c968d9105296cae83d51b1c58def",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "color": {
                                        "type": "string",
                                        "example": "#000000"
                                    },
                                    "eye_color": {
                                        "type": "string",
                                        "example": "#ff0000"
                                    },
                                    "logo_url": {
                                        "type": "string",
                                        "format": "uri"
                                    },
                                    "style": {
                                        "type": "string",
                                        "enum": [
                                            "square",
                                            "round",
                                            "dot"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Branding updated"
                    },
                    "404": {
                        "description": "Link not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}